fix(pageBlocks): rename DefaultLayoutProvider to LayoutProvider
