+++ packages/xprog-prensa/providers/LayoutProvider/provider.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
 
 import * as data from './data'
-import { DefaultLayoutProviderProps } from './props'
+import { LayoutProviderProps } from './props'
 
-export const DefaultLayoutProvider = (props: DefaultLayoutProviderProps) => {
+export const LayoutProvider = (props: LayoutProviderProps) => {
   const { children, ...otherProps } = props
   return (
-    <div className={'defaultLayoutProvider'}>
+    <div className={'LayoutProvider'}>
       {React.Children.map(children, (item: any) =>
         React.cloneElement(item, { ...data.pageblockConfig, ...data, ...otherProps })
       )}
